# src/main/resources/布灵4.0.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "布灵",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/knowledge-bases/{kbId}/resources/tree": {
      "get": {
        "summary": "获取知识库下的资源目录树",
        "deprecated": false,
        "description": "要么是点进知识库,要么是点到知识库的文档,\n使用此接口,增加知识库view_count",
        "operationId": "getResourceTree",
        "tags": [
          "资源 (Resource)"
        ],
        "parameters": [
          {
            "name": "kbId",
            "in": "path",
            "description": "知识库ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultResourceTree"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/knowledge-bases/{kbId}/resources": {
      "post": {
        "summary": "在知识库中新建资源",
        "deprecated": false,
        "description": "数据库中新建一个资源记录,不实际新建资源内容",
        "operationId": "createResource",
        "tags": [
          "资源 (Resource)"
        ],
        "parameters": [
          {
            "name": "kbId",
            "in": "path",
            "description": "目标知识库ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourceCreate"
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/resources/{resId}": {
      "get": {
        "summary": "查看单个资源的完整信息",
        "deprecated": false,
        "description": "会增加访问量",
        "operationId": "getResourceById",
        "tags": [
          "资源 (Resource)"
        ],
        "parameters": [
          {
            "name": "resId",
            "in": "path",
            "description": "资源ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultResource"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "put": {
        "summary": "更新资源内容或标题",
        "deprecated": false,
        "description": "",
        "operationId": "updateResource",
        "tags": [
          "资源 (Resource)"
        ],
        "parameters": [
          {
            "name": "resId",
            "in": "path",
            "description": "资源ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "删除资源 (逻辑删除)",
        "deprecated": false,
        "description": "删除资源及其子资源",
        "operationId": "deleteResource",
        "tags": [
          "资源 (Resource)"
        ],
        "parameters": [
          {
            "name": "resId",
            "in": "path",
            "description": "资源ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/resources/{resId}/visibility": {
      "patch": {
        "summary": "更新资源可见性",
        "deprecated": false,
        "description": "",
        "operationId": "updateResourceVisibility",
        "tags": [
          "资源 (Resource)"
        ],
        "parameters": [
          {
            "name": "resId",
            "in": "path",
            "description": "资源ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "visibility": {
                    "type": "integer",
                    "description": "可见性: 0-私密, 1-公开",
                    "examples": [
                      1
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/resources/{resId}/status": {
      "patch": {
        "summary": "更新资源上架/下架状态",
        "deprecated": false,
        "description": "",
        "operationId": "updateResourceStatus",
        "tags": [
          "资源 (Resource)"
        ],
        "parameters": [
          {
            "name": "resId",
            "in": "path",
            "description": "资源ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "integer",
                    "description": "状态: 0-下架, 1-上架",
                    "examples": [
                      1
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/resources/{resId}/rename": {
      "patch": {
        "summary": "重命名资源",
        "deprecated": false,
        "description": "",
        "operationId": "renameResource",
        "tags": [
          "资源 (Resource)"
        ],
        "parameters": [
          {
            "name": "resId",
            "in": "path",
            "description": "资源ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "examples": [
                      "新的资源标题"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/resources/{resId}/move": {
      "post": {
        "summary": "移动资源或目录",
        "deprecated": false,
        "description": "",
        "operationId": "moveResource",
        "tags": [
          "资源 (Resource)"
        ],
        "parameters": [
          {
            "name": "resId",
            "in": "path",
            "description": "要移动的资源或目录ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "targetKbId": {
                    "type": "string",
                    "format": "int64",
                    "description": "目标知识库ID"
                  },
                  "targetPreId": {
                    "type": "string",
                    "format": "int64",
                    "description": "目标父级目录ID (资源ID)",
                    "nullable": true
                  }
                },
                "required": [
                  "targetKbId"
                ]
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "移动成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/resources/{resId}/copy": {
      "post": {
        "summary": "复制资源",
        "deprecated": false,
        "description": "",
        "operationId": "copyResource",
        "tags": [
          "资源 (Resource)"
        ],
        "parameters": [
          {
            "name": "resId",
            "in": "path",
            "description": "要复制的资源ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourceMoveCopy"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "复制成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/resources/{resId}/copy-tree": {
      "post": {
        "summary": "复制目录(及目录下所有子资源)",
        "deprecated": false,
        "description": "拷贝一份目录与目录下的子资源",
        "operationId": "copyResourceTree",
        "tags": [
          "资源 (Resource)"
        ],
        "parameters": [
          {
            "name": "resId",
            "in": "path",
            "description": "要复制的目录ID (本身也是一个资源)",
            "required": true,
            "example": "",
            "schema": {
              "type": "string",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourceMoveCopy"
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "复制成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/resources/{resId}/share": {
      "post": {
        "summary": "生成资源分享链接",
        "deprecated": false,
        "description": "",
        "operationId": "generateResourceShareLink",
        "tags": [
          "分享 (Share)"
        ],
        "parameters": [
          {
            "name": "resId",
            "in": "path",
            "description": "资源ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "生成成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "examples": [
                        200
                      ]
                    },
                    "message": {
                      "type": "string",
                      "examples": [
                        "Success"
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "share_url": {
                          "type": "string",
                          "examples": [
                            "http://localhost:3000/share/kbs_sh_xxxxx/res_sh_yyyyy"
                          ]
                        }
                      }
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/share/{kbShareId}/{resShareId}": {
      "get": {
        "summary": "访问分享链接查看资源",
        "deprecated": false,
        "description": "这是一个公开接口, 无需认证",
        "operationId": "accessSharedResource",
        "tags": [
          "分享 (Share)"
        ],
        "parameters": [
          {
            "name": "kbShareId",
            "in": "path",
            "description": "知识库的分享ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resShareId",
            "in": "path",
            "description": "资源的分享ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultResource"
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "分享的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/recycle-bin/resources/{resId}": {
      "post": {
        "summary": "恢复资源",
        "deprecated": false,
        "description": "恢复至原知识库下根节点",
        "operationId": "restoreResource",
        "tags": [
          "回收站 (Recycle Bin)"
        ],
        "parameters": [
          {
            "name": "resId",
            "in": "path",
            "description": "要恢复的资源ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "恢复成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "彻底删除资源",
        "deprecated": true,
        "description": "",
        "operationId": "permanentlyDeleteResource",
        "tags": [
          "回收站 (Recycle Bin)"
        ],
        "parameters": [
          {
            "name": "resId",
            "in": "path",
            "description": "要彻底删除的资源ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/explore/articles": {
      "get": {
        "summary": "获取推荐文章列表",
        "deprecated": false,
        "description": "在首页或“逛逛”页面分页获取公开的推荐文章。",
        "operationId": "listExploreArticles",
        "tags": [
          "探索与推荐 (Explore & Recommendation)"
        ],
        "parameters": [
          {
            "name": "sortBy",
            "in": "query",
            "description": "排序方式",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "hot",
                "new"
              ],
              "default": "hot"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "每页数量",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultArticleSummaryList"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/submissions": {
      "post": {
        "summary": "提交投稿（申请推荐）",
        "deprecated": false,
        "description": "用户将自己知识库中的某一篇文章提交申请，希望被推荐到“逛逛”模块。",
        "operationId": "createSubmission",
        "tags": [
          "探索与推荐 (Explore & Recommendation)"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmissionCreateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "请求已被接受，等待审核",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Result": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码 (200:成功, 其他:失败)",
            "examples": [
              200
            ]
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "examples": [
              "Success"
            ]
          },
          "data": {
            "type": "object",
            "description": "响应数据",
            "properties": {},
            "nullable": true
          }
        }
      },
...
      "ResourceCreate": {
        "type": "object",
        "properties": {
          "preId": {
            "type": "string",
            "format": "int64",
            "description": "父级资源ID, 如果是根目录则为null",
            "examples": [
              "12345678901234567"
            ],
            "nullable": true
          },
          "title": {
            "type": "string",
            "description": "资源标题",
            "default": "无标题资源",
            "examples": [
              "我的第一篇资源"
            ]
          },
          "type": {
            "type": "string",
            "description": "资源类型, e.g., markdown, doc, sheet",
            "examples": [
              "markdown"
            ]
          }
        }
      },
...
      "ResultResourceTree": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Result"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ResourceNode"
                }
              }
            }
          }
        ]
      },
...
    },
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}